+++ Server/Application/Application+Queries.m
@@ -10,6 +10,7 @@
 #import "XCAXClient_iOS.h"
 #import "XCUIElement.h"
 #import "JSONUtils.h"
+#import "XCApplicationQuery.h"
 
 @implementation Application (Queries)
 static NSArray <NSString *> *identifierProperties;
@@ -38,8 +39,8 @@ static NSArray <NSString *> *identifierProperties;
 }
 
 + (NSDictionary *)snapshotTree:(XCElementSnapshot *)snapshot {
-    NSMutableDictionary *json = [JSONUtils snapshotToJSON:snapshot];
-    
+    NSMutableDictionary *json = [JSONUtils snapshotOrElementToJSON:snapshot];
+
     if (snapshot.children.count) {
         NSMutableArray *children = [NSMutableArray array];
         for (XCElementSnapshot *child in snapshot.children) {
@@ -87,7 +88,7 @@ static NSArray <NSString *> *identifierProperties;
     NSPredicate *predicate = [NSPredicate predicateWithFormat:predString];
     XCUIElementQuery *query = [[[self currentApplication] descendantsMatchingType:XCUIElementTypeAny]
                                matchingPredicate:predicate];
-    
+
     return [query allElementsBoundByIndex];
 }
 
@@ -96,7 +97,7 @@ static NSArray <NSString *> *identifierProperties;
     NSArray *childElements = [self elementsWithAnyOfTheseProperties:properties
                                                        equalToValue:value];
     NSMutableArray *ret = [NSMutableArray array];
-    
+
     //TODO: Should we check the application itself?
     for (XCUIElement *element in childElements) {
         [ret addObject:[JSONUtils elementToJSON:element]];
@@ -107,7 +108,7 @@ static NSArray <NSString *> *identifierProperties;
 + (NSArray <NSDictionary *> *)jsonForElementsWithID:(NSString *)identifier {
     NSArray *elements = [self elementsWithIdentifier:identifier];
     NSMutableArray *ret = [NSMutableArray array];
-    
+
     //TODO: Should we check the application itself?
     for (XCUIElement *element in elements) {
         [ret addObject:[JSONUtils elementToJSON:element]];
@@ -118,7 +119,7 @@ static NSArray <NSString *> *identifierProperties;
 + (NSArray <NSDictionary *> *)jsonForElementsWithType:(NSString *)type {
     NSArray *elements = [self elementsWithType:type];
     NSMutableArray *ret = [NSMutableArray array];
-    
+
     //TODO: Should we check the application itself?
     for (XCUIElement *element in elements) {
         [ret addObject:[JSONUtils elementToJSON:element]];