+++ Server/Application/SpringBoard.m
@@ -13,8 +13,11 @@
 #import "XCUIElement.h"
 #import "XCElementSnapshot.h"
 #import "GestureFactory.h"
+#import "XCUICoordinate.h"
 #import "XCUIElement+WebDriverAttributes.h"
 #import "CBXException.h"
+#import <UIKit/UIKit.h>
+#import "XCApplicationQuery.h"
 
 typedef enum : NSUInteger {
     SpringBoardAlertHandlerIgnoringAlerts = 0,
@@ -25,9 +28,12 @@ typedef enum : NSUInteger {
 
 @interface SpringBoard ()
 
+- (BOOL)UIApplication_isSpringBoardShowingAnAlert;
 - (BOOL)shouldDismissAlertsAutomatically;
 - (SpringBoardAlertHandlerResult)handleAlert;
-- (void)tapAlertButton:(XCUIElement *)button;
+- (BOOL)tapAlertButton:(XCUIElement *)alertButton;
+- (CGPoint)hitPointForAlertButton:(XCUIElement *)alertButton;
+- (CGPoint)pointByTranslatingPoint:(CGPoint)point;
 
 @end
 
@@ -55,15 +61,48 @@ typedef enum : NSUInteger {
     return _springBoard;
 }
 
+- (BOOL)UIApplication_isSpringBoardShowingAnAlert {
+    SEL selector = NSSelectorFromString(@"_isSpringBoardShowingAnAlert");
+    if (![[UIApplication sharedApplication] respondsToSelector:selector]) {
+        DDLogDebug(@"UIApplication does not respond to %@; returning YES to force XCUIElementQuery",
+              NSStringFromSelector(selector));
+        return YES;
+    }
+
+    NSMethodSignature *signature;
+    signature = [[UIApplication class] instanceMethodSignatureForSelector:selector];
+    NSInvocation *invocation;
+
+    invocation = [NSInvocation invocationWithMethodSignature:signature];
+    invocation.target = [UIApplication sharedApplication];
+    invocation.selector = selector;
+
+    [invocation invoke];
+
+    BOOL alertShowing = NO;
+    char ref;
+    [invocation getReturnValue:(void **) &ref];
+    if (ref == (BOOL)1) {
+        alertShowing = YES;
+    }
+
+    return alertShowing;
+}
+
 - (XCUIElement *)queryForAlert {
     @synchronized (self) {
         XCUIElement *alert = nil;
 
-        XCUIElementQuery *query = [self descendantsMatchingType:XCUIElementTypeAlert];
-        NSArray <XCUIElement *> *elements = [query allElementsBoundByIndex];
+        [self _waitForQuiescence];
 
-        if ([elements count] != 0) {
-            alert = elements[0];
+        if([self UIApplication_isSpringBoardShowingAnAlert]) {
+
+            XCUIElementQuery *query = [self descendantsMatchingType:XCUIElementTypeAlert];
+            NSArray <XCUIElement *> *elements = [query allElementsBoundByIndex];
+
+            if ([elements count] != 0) {
+                alert = elements[0];
+            }
         }
         return alert;
     }
@@ -201,15 +240,75 @@ typedef enum : NSUInteger {
         return SpringBoardAlertHandlerNoAlert;
     }
 
-    [self tapAlertButton:button];
+    BOOL success = [self tapAlertButton:button];
+
+    if (!success) {
+        return SpringBoardAlertHandlerNoAlert;
+    }
 
     return SpringBoardAlertHandlerDismissedAlert;
 }
 
-- (void)tapAlertButton:(XCUIElement *)button {
+- (SpringBoardDismissAlertResult)dismissAlertByTappingButtonWithTitle:(NSString *)title {
     @synchronized (self) {
-        [button tap];
+        XCUIElement *alert = [self queryForAlert];
 
+        if (!alert) {
+            return SpringBoardDismissAlertNoAlert;
+        } else {
+            XCUIElement *button = alert.buttons[title];
+            [button resolve];
+
+            if (!button || !button.exists) {
+                return SpringBoardDismissAlertNoMatchingButton;
+            }
+
+            BOOL success = [self tapAlertButton:button];
+
+            SpringBoardDismissAlertResult result;
+            if (success) {
+                result = SpringBoardDismissAlertDismissedAlert;
+            } else {
+                result = SpringBoardDismissAlertDismissTouchFailed;
+            }
+
+            return result;
+        }
+    }
+}
+
+- (BOOL)tapAlertButton:(XCUIElement *)alertButton {
+    @synchronized (self) {
+        [alertButton resolve];
+        CGPoint hitPoint = [self hitPointForAlertButton:alertButton];
+
+        if (hitPoint.x < 0 || hitPoint.y < 0) {
+           return NO;
+        }
+
+        NSDictionary *body =
+        @{
+          @"gesture" : @"touch",
+          @"options" : @{},
+          @"specifiers" : @{@"coordinate" :
+                                @{ @"x" : @(hitPoint.x),
+                                   @"y" : @(hitPoint.y)
+                                   }
+                            }
+          };
+
+        __block BOOL success = YES;
+        [GestureFactory executeGestureWithJSON:body completion:^(NSError *error) {
+            if (error) {
+                DDLogError(@"Error dismissing alert: %@", error.localizedDescription);
+                success = NO;
+            }
+        }];
+
+        // Let the main RunLoop progress before executing more queries or
+        // gestures.  It is possible that a failed touch will succeed on the
+        // next pass.
+        //
         // There is one alert workflow that is very problematic:
         //
         // PhotoRoll
@@ -227,7 +326,7 @@ typedef enum : NSUInteger {
         // of the Cancel touch.  Sleeping after the dismiss definitely
         // reduced the frequency of crashes - they still happened.
         //
-        // The AUT crash was caused by IImagePickerViewController which has a
+        // The AUT crash was caused by UIImagePickerViewController which has a
         // history of crashing in situations like this.
         //
         // After days device and simulator testing, I settled on 1.0 second.
@@ -241,28 +340,54 @@ typedef enum : NSUInteger {
         // We prefer stability over speed.
         CFTimeInterval interval = 1.0;
         CFRunLoopRunInMode(kCFRunLoopDefaultMode, interval, false);
-    }
+
+        return success;
+   }
 }
 
-- (SpringBoardDismissAlertResult)dismissAlertByTappingButtonWithTitle:(NSString *)title {
-    @synchronized (self) {
-        XCUIElement *alert = [self queryForAlert];
+- (CGPoint)hitPointForAlertButton:(XCUIElement *)alertButton {
+    [alertButton resolve];
 
-        if (!alert) {
-            return SpringBoardDismissAlertNoAlert;
-        } else {
-            XCUIElement *button = alert.buttons[title];
-            [button resolve];
+    XCElementSnapshot *snapshot = alertButton.lastSnapshot;
 
-            if (!button.exists) {
-                return SpringBoardDismissAlertNoMatchingButton;
-            }
+    NSValue *hitPointValue = snapshot.suggestedHitpoints.firstObject;
 
-            [self tapAlertButton:button];
-
-            return SpringBoardDismissAlertDismissedAlert;
-        }
+    CGPoint hitPoint;
+    if (!hitPointValue) {
+        XCUICoordinate *coordinate;
+        coordinate = [alertButton coordinateWithNormalizedOffset:CGVectorMake(0.5, 0.5)];
+        hitPoint = coordinate.screenPoint;
+    } else {
+        hitPoint = hitPointValue.CGPointValue;
     }
+
+    return [self pointByTranslatingPoint:hitPoint];
+}
+
+- (CGPoint)pointByTranslatingPoint:(CGPoint)point {
+    CGPoint translated = point;
+    CGSize appSize = self.frame.size;
+    UIInterfaceOrientation orientation = self.interfaceOrientation;
+
+    switch (orientation) {
+        case UIInterfaceOrientationPortraitUpsideDown: {
+            translated = CGPointMake(appSize.width - point.x,
+                                     appSize.height - point.y);
+            break;
+        }
+        case UIInterfaceOrientationLandscapeLeft: {
+            translated = CGPointMake(appSize.height - point.y, point.x);
+            break;
+        }
+        case UIInterfaceOrientationLandscapeRight: {
+            translated = CGPointMake(point.y, appSize.width - point.x);
+            break;
+        }
+
+        default: { break; }
+    }
+
+    return translated;
 }
 
 @end