+++ Server/Routes/QueryRoutes.m
@@ -25,16 +25,16 @@
                  QueryConfiguration *queryConfig = [QueryConfigurationFactory configWithJSON:body
                                                                                    validator:[Query validator]];
                  Query *query = [QueryFactory queryWithQueryConfiguration:queryConfig];
-                 
+
                  NSArray <XCUIElement *> *elements = [query execute];
-                 
+
                  /*
                     Format and return the results
                   */
                  NSMutableArray *results = [NSMutableArray arrayWithCapacity:elements.count];
                  for (XCUIElement *el in elements) {
                      [Application cacheElement:el];
-                     NSDictionary *json = [JSONUtils snapshotToJSON:el];
+                     NSDictionary *json = [JSONUtils snapshotOrElementToJSON:el];
                      [results addObject:json];
                  }
                  [response respondWithJSON:@{@"result" : results}];
@@ -58,19 +58,19 @@
                  NSString *identifier = request.params[CBX_IDENTIFIER_KEY];
                  [response respondWithJSON:[Application jsonForElementsWithID:identifier]];
              }],
-             
+
              [CBXRoute get:endpoint(@"/query/test_id/:test_id", 1.0) withBlock:^(RouteRequest *request, NSDictionary *data, RouteResponse *response) {
                  NSNumber *identifier = @([request.params[CBX_TEST_ID_KEY] integerValue]);
                  XCUIElement *el = [Application cachedElementOrThrow:identifier];
                  [response respondWithJSON:[JSONUtils elementToJSON:el]];
              }],
-             
+
              [CBXRoute get:endpoint(@"/query/type/:type", 1.0) withBlock:^(RouteRequest *request, NSDictionary *data, RouteResponse *response) {
                  NSString *type = request.params[CBX_TYPE_KEY];
                  [response respondWithJSON:[Application jsonForElementsWithType:type]];
              }],
-             
-             
+
+
              ];
 }
 @end