+++ Server/XCTest/XCElementSnapshot.h
@@ -4,8 +4,7 @@
 //     class-dump is Copyright (C) 1997-1998, 2000-2001, 2004-2013 by Steve Nygard.
 //
 
-//#import <XCTWebDriverAgentLib/CDStructures.h>
-
+#import "CDStructures.h"
 #import <XCTest/XCUIElementAttributes.h>
 
 @class NSArray, NSDictionary, NSString, XCAccessibilityElement, XCUIApplication;
@@ -71,9 +70,9 @@
 - (BOOL)_frameFuzzyMatchesElement:(id)arg1;
 - (BOOL)_fuzzyMatchesElement:(id)arg1;
 - (BOOL)_matchesElement:(id)arg1;
-- (id)_allDescendants;
-//- (BOOL)hasDescendantMatchingFilter:(CDUnknownBlockType)arg1;
-- (id)descendantsByFilteringWithBlock:(BOOL(^)(XCElementSnapshot *snapshot))block;
+- (NSArray<XCElementSnapshot *> *)_allDescendants;
+- (BOOL)hasDescendantMatchingFilter:(CDUnknownBlockType)arg1;
+- (NSArray<XCElementSnapshot *> *)descendantsByFilteringWithBlock:(BOOL(^)(XCElementSnapshot *snapshot))block;
 - (id)elementSnapshotMatchingAccessibilityElement:(id)arg1;
 - (void)enumerateDescendantsUsingBlock:(void(^)(XCElementSnapshot *snapshot))block;
 - (id)recursiveDescriptionWithIndent:(id)arg1 includeAccessibilityElement:(BOOL)arg2;