+++ cucumber/bin/test/jmoody/sierra.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 set -e
-CUCUMBER_ARGS="--format pretty --tags @keyboard"
+CUCUMBER_ARGS="--format pretty --tags ~@tree"
 
 if [ "${1}" = "device" ]; then
   TEST_SIM=0
@@ -54,18 +54,18 @@ banner "Xcode 8.0"
 xcrun xcodebuild -version
 
 if [ $TEST_SIM -eq 1 ]; then
-  info "Default Simulator"
+  info "Xcode 8.0 Default Simulator"
   bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 9.3 Simulator"
+  info "Xcode 8.0 iOS 9.3 Simulator"
   DEVICE_TARGET=7D25AA5B-F939-4BC8-AD5D-6280E71DAE0B \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 9.3 Simulator iPhone 5c"
+  info "Xcode 8.0 iOS 9.3 Simulator iPhone 5c"
   DEVICE_TARGET=3BFEF174-BBF3-48C0-B860-7898250F2254 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.0 Simulator iPhone 5c"
+  info "Xcode 8.0 iOS 10.0 Simulator iPhone 5c"
   DEVICE_TARGET=42EF5066-9946-498C-A684-9C8AF702DEC3 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 fi
@@ -75,7 +75,7 @@ if [ $TEST_DEVICE -eq 1 ]; then
 
   for device in "${devices[@]}"
   do
-    info "Device: $device"
+    info "Xcode 8.0 with device: $device"
 
     DEVICE_TARGET=$device \
       APP=sh.calaba.TestApp \
@@ -90,26 +90,26 @@ banner "Xcode 8.1"
 xcrun xcodebuild -version
 
 if [ $TEST_SIM -eq 1 ]; then
-  info "Default Simulator"
+  info "Xcode 8.1 Default Simulator"
   bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.0 Simulator"
+  info "Xcode 8.1 iOS 10.0 Simulator"
   DEVICE_TARGET=B6D21623-32F6-44F5-90ED-7E520DC94A69 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 9.3 Simlator"
+  info "Xcode 8.1 iOS 9.3 Simlator"
   DEVICE_TARGET=6DD279B3-7FED-4902-96F4-618D55540ADC \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 9.3 Simulator iPhone 5c"
+  info "Xcode 8.1 iOS 9.3 Simulator iPhone 5c"
   DEVICE_TARGET=3BFEF174-BBF3-48C0-B860-7898250F2254 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.0 Simulator iPhone 5c"
+  info "Xcode 8.1 iOS 10.0 Simulator iPhone 5c"
   DEVICE_TARGET=42EF5066-9946-498C-A684-9C8AF702DEC3 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.1 Simulator iPhone 5c"
+  info "Xcode 8.1 iOS 10.1 Simulator iPhone 5c"
   DEVICE_TARGET=659F0339-A6A9-460E-BFDD-30E7F12F7E7C \
     bundle exec cucumber ${CUCUMBER_ARGS}
 fi
@@ -119,58 +119,58 @@ if [ $TEST_DEVICE -eq 1 ]; then
 
   for device in "${devices[@]}"
   do
-    info "Device: $device"
+    info "Xcode 8.1 with device: $device"
     DEVICE_TARGET=$device \
       APP=sh.calaba.TestApp \
       bundle exec cucumber ${CUCUMBER_ARGS}
   done
 fi
 
-# Xcode 8.2 beta 2
-export DEVELOPER_DIR=/Xcode/8.2/Xcode-beta.app/Contents/Developer
+# Xcode 8.2.1
+export DEVELOPER_DIR=/Xcode/8.2.1/Xcode.app/Contents/Developer
 
-banner "Xcode 8.2"
+banner "Xcode 8.2.1"
 xcrun xcodebuild -version
 
 if [ $TEST_SIM -eq 1 ]; then
-  info "Default Simulator"
+  info "Xcode 8.2.1 Default Simulator"
   bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.1 Simulator"
+  info "Xcode 8.2.1 iOS 10.1 Simulator"
   DEVICE_TARGET=E160BC4E-CAD5-4680-9C42-D510FB4BB052 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.0 Simlator"
-  DEVICE_TARGET=6DD279B3-7FED-4902-96F4-618D55540ADC \
-    bundle exec cucumber ${CUCUMBER_ARGS}
-
-  info "iOS 9.3 Simulator"
+  info "Xcode 8.2.1 iOS 10.0 Simlator"
   DEVICE_TARGET=B6D21623-32F6-44F5-90ED-7E520DC94A69 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 9.3 Simulator iPhone 5c"
+  info "Xcode 8.2.1 iOS 9.3 Simulator"
+  DEVICE_TARGET=6DD279B3-7FED-4902-96F4-618D55540ADC \
+    bundle exec cucumber ${CUCUMBER_ARGS}
+
+  info "Xcode 8.2.1 iOS 9.3 Simulator iPhone 5c"
   DEVICE_TARGET=3BFEF174-BBF3-48C0-B860-7898250F2254 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.0 Simulator iPhone 5c"
+  info "Xcode 8.2.1 iOS 10.0 Simulator iPhone 5c"
   DEVICE_TARGET=42EF5066-9946-498C-A684-9C8AF702DEC3 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.1 Simulator iPhone 5c"
+  info "Xcode 8.2.1 iOS 10.1 Simulator iPhone 5c"
   DEVICE_TARGET=659F0339-A6A9-460E-BFDD-30E7F12F7E7C \
     bundle exec cucumber ${CUCUMBER_ARGS}
 
-  info "iOS 10.2 Simulator iPhone 5c"
+  info "Xcode 8.2.1 iOS 10.2 Simulator iPhone 5c"
   DEVICE_TARGET=83B48D5E-A97B-4CE0-8B43-AD8B841303A7 \
     bundle exec cucumber ${CUCUMBER_ARGS}
 fi
 
 if [ $TEST_DEVICE -eq 1 ]; then
-  declare -a devices=("mercury" "wolf" "uranus" "pegasi" "denis" "hat")
+  declare -a devices=("mercury" "wolf" "uranus" "pegasi" "denis")
 
   for device in "${devices[@]}"
   do
-    info "Device: $device"
+    info "Xcode 8.2.1 with device: $device"
     DEVICE_TARGET=$device \
       APP=sh.calaba.TestApp \
       bundle exec cucumber ${CUCUMBER_ARGS}