+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### 1.0.5
+
+* Increase touch duration to 0.2 to match XCUITest #198
+* Use CocoaLumberjack for logging #197
+* Update CocoaLumberjack to 3.0.0 #196
+* TestApp: update to calabash.framework 0.20.4 #194
+* Improve hitpoint and visibility calculation #193
+* Fix text entry for i386/armv7 devices #192
+* Dismiss SpringBoard alerts by touching alert buttons with coordinates #191
+* SpringBoard: ask UIApplication for SpringBoard alert before making an
+  expensive XCUITest query #190
+* POST /session raises when app is not installed and when app fails to
+  launch #189
+
 ### 1.0.4
 
 DeviceAgent can dismiss SpringBoard alerts in any orientation.